add jwt support.

+++ app/views/account.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python
 # -*- encoding: utf-8 -*-
 from flask import (Blueprint, request)
-from flask_restful import Api
+from flask_restful import (Api, abort)
 from webargs.flaskparser import (use_args, parser)
+from flask_jwt_extended import jwt_required
 
 from app.models import Account
 from extensions import db
@@ -12,6 +13,7 @@ from app.libs.error import error
 
 
 class AccountsResource(BaseResource):
+    @jwt_required
     @use_args(page_args, locations=('query',))
     def get(self, args):
         accounts = Account.query.order_by(Account.id.desc()).paginate(
@@ -20,6 +22,7 @@ class AccountsResource(BaseResource):
 
         return self.paginate(result.data, accounts.per_page, accounts.total)
 
+    @jwt_required
     @use_args(account_schema)
     def post(self, args):
         account = Account(**args)
@@ -34,11 +37,13 @@ class AccountsResource(BaseResource):
 
 
 class AccountResource(BaseResource):
+    @jwt_required
     @BaseResource.check_record(Account)
     def get(self, account_id):
         result = account_schema.dump(self.record)
         return result.data
 
+    @jwt_required
     @BaseResource.check_record(Account)
     def put(self, account_id):
         args = parser.parse(account_schema, request)
@@ -54,12 +59,15 @@ class AccountResource(BaseResource):
 
         return result.data
 
+    @jwt_required
     @BaseResource.check_record(Account)
     def delete(self, account_id):
-        db.session.delete(self.record)
-        db.session.commit()
-        return None, 204
+        if self.current_user and self.current_user.id == account_id:
+            db.session.delete(self.record)
+            db.session.commit()
+            return None, 204
 
+        return abort(403, message='only can be deleted by himself')
 
 account_bp = Blueprint('account', __name__)
 account_api = Api(account_bp, prefix='/api/v1')
